testcase updates5

+++ br/pkg/task/backup.go
@@ -631,7 +631,7 @@ func RunBackup(c context.Context, g glue.Glue, cmdName string, cfg *BackupConfig
 	progressCount := uint64(0)
 	progressCallBack := func() {
 		updateCh.Inc()
-		failpoint.Inject("progress-call-back", func(v failpoint.Value) {
+		if v, _err_ := failpoint.Eval(_curpkg_("progress-call-back")); _err_ == nil {
 			log.Info("failpoint progress-call-back injected")
 			atomic.AddUint64(&progressCount, 1)
 			if fileName, ok := v.(string); ok {
@@ -645,7 +645,7 @@ func RunBackup(c context.Context, g glue.Glue, cmdName string, cfg *BackupConfig
 					log.Warn("failed to write data to file", zap.Error(err))
 				}
 			}
-		})
+		}
 	}
 
 	if cfg.UseCheckpoint {
@@ -668,7 +668,7 @@ func RunBackup(c context.Context, g glue.Glue, cmdName string, cfg *BackupConfig
 		}()
 	}
 
-	failpoint.Inject("s3-outage-during-writing-file", func(v failpoint.Value) {
+	if v, _err_ := failpoint.Eval(_curpkg_("s3-outage-during-writing-file")); _err_ == nil {
 		log.Info("failpoint s3-outage-during-writing-file injected, " +
 			"process will sleep for 5s and notify the shell to kill s3 service.")
 		if sigFile, ok := v.(string); ok {
@@ -681,7 +681,7 @@ func RunBackup(c context.Context, g glue.Glue, cmdName string, cfg *BackupConfig
 			}
 		}
 		time.Sleep(5 * time.Second)
-	})
+	}
 
 	metawriter.StartWriteMetasAsync(ctx, metautil.AppendDataFile)
 	err = client.BackupRanges(ctx, ranges, req, uint(cfg.Concurrency), cfg.ReplicaReadLabel, metawriter, progressCallBack)
